+++ backend/tests/post-service-test.js
@@ -111,8 +111,7 @@ let posts = []; //fake database
 
 describe('Post services and model', function () {
 
-    //clear out the posts array
-    beforeEach(async () => {
+    before(async () => {
         if (process.env.TEST_TYPE === 'INTEGRATION') {
             mongoose
             .connect(process.env.MONGODB_CONNECTION, {
@@ -120,17 +119,22 @@ describe('Post services and model', function () {
                 useUnifiedTopology: true,})
                 .then(() => {console.log("Success to connect mongodb");})
                 .catch(() => {console.log("Fail to connect mongodb")});
+        }
+        else {
+            setFakeDatabase();
+        }
+    })
 
+    beforeEach(async () => {
+        if (process.env.TEST_TYPE === 'INTEGRATION') {
             await Post.deleteMany({});
         }
         else {
             posts = [];
-            setFakeDatabase();
         }
     });
 
-    //get rid of all stubs
-    afterEach(async () => {
+    after(async () => {
         if (process.env.TEST_TYPE === 'INTEGRATION') {
             await mongoose.disconnect();
         }
